got the analytics hook working
