+++ src/hooks/index.js
@@ -34,7 +34,8 @@ export { default as useDPI } from './useDPI';
 export { default as useEffectDebugger } from './useEffectDebugger';
 export { default as useEmailRainbow } from './useEmailRainbow';
 export { default as useENSPendingRegistrations } from './useENSPendingRegistrations';
-export { default as useENSProfile, trackProfile } from './useENSProfile';
+export { default as useENSProfile } from './useENSProfile';
+export { default as useTrackENSProfile } from './useTrackENSProfile';
 export { default as useENSRecordDisplayProperties } from './useENSRecordDisplayProperties';
 export { default as useENSRegistration } from './useENSRegistration';
 export { default as useENSRegistrationActionHandler } from './useENSRegistrationActionHandler';
