+++ src/screens/WalletScreen.js
@@ -8,6 +8,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import { OpacityToggler } from '../components/animations';
 import { AssetList } from '../components/asset-list';
 import { ExchangeFab, FabWrapper, SendFab } from '../components/fab';
+import useExperimentalFlag, {
+  PROFILES,
+} from '@rainbow-me/config/experimentalHooks';
 import {
   DiscoverHeaderButton,
   Header,
@@ -20,7 +23,7 @@ import {
   useAccountEmptyState,
   useAccountSettings,
   useCoinListEdited,
-  trackProfile,
+  useTrackENSProfile,
   useInitializeDiscoverData,
   useInitializeWallet,
   useLoadGlobalLateData,
@@ -64,6 +67,7 @@ export default function WalletScreen() {
   const scrollViewTracker = useValue(0);
   const { isReadOnlyWallet } = useWallets();
   const { isEmpty: isAccountEmpty } = useAccountEmptyState();
+  const { trackENSProfile } = useTrackENSProfile();
   const { network } = useAccountSettings();
   const { userAccounts } = useUserAccounts();
   const { portfolios, trackPortfolios } = usePortfolios();
@@ -88,6 +92,8 @@ export default function WalletScreen() {
       assetsSocket,
     })
   );
+
+  const profilesEnabled = useExperimentalFlag(PROFILES);
 
   useEffect(() => {
     const fetchAndResetFetchSavings = async () => {
@@ -134,9 +140,10 @@ export default function WalletScreen() {
   ]);
 
   useEffect(() => {
-    console.log('HELLLOOOO');
-    trackProfile();
-  }, [trackProfile]);
+    if (profilesEnabled) {
+      trackENSProfile();
+    }
+  }, [profilesEnabled, trackENSProfile]);
 
   useEffect(() => {
     if (
